+++ NEWS.md
@@ -1,4 +1,12 @@
 
+# tfrmtbuilder 0.0.2
+
+* Data Mapping improvements:
+  - Persistent mapping selections when additional inputs are added
+  - Informing user when the "save" button has yet to be pressed
+* Mock mode toggle in nav bar rather than load tab
+
+
 # tfrmtbuilder 0.0.1
 
 First release